+++ src/store/src/Bridge/Azure/SearchStore.php
@@ -15,14 +15,14 @@ use Symfony\AI\Platform\Vector\NullVector;
 use Symfony\AI\Platform\Vector\Vector;
 use Symfony\AI\Store\Document\Metadata;
 use Symfony\AI\Store\Document\VectorDocument;
-use Symfony\AI\Store\VectorStoreInterface;
+use Symfony\AI\Store\StoreInterface;
 use Symfony\Component\Uid\Uuid;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
  * @author Christopher Hertel <mail@christopher-hertel.de>
  */
-final readonly class SearchStore implements VectorStoreInterface
+final readonly class SearchStore implements StoreInterface
 {
     /**
      * @param string $vectorFieldName The name of the field int the index that contains the vector