+++ src/store/src/Bridge/SurrealDb/Store.php
@@ -18,14 +18,14 @@ use Symfony\AI\Store\Document\VectorDocument;
 use Symfony\AI\Store\Exception\InvalidArgumentException;
 use Symfony\AI\Store\Exception\RuntimeException;
 use Symfony\AI\Store\InitializableStoreInterface;
-use Symfony\AI\Store\VectorStoreInterface;
+use Symfony\AI\Store\StoreInterface;
 use Symfony\Component\Uid\Uuid;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
  * @author Guillaume Loulier <personal@guillaumeloulier.fr>
  */
-final class Store implements InitializableStoreInterface, VectorStoreInterface
+final class Store implements InitializableStoreInterface, StoreInterface
 {
     private string $authenticationToken = '';